pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-secret-validator"
version = "0.1.0"
description = "LLM-based dynamic secret detection and validation engine"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.30.6",
  "pydantic==2.9.2",
  "requests==2.32.3",
  "google-genai==0.3.0"
]

[project.scripts]
mcp-scan = "cli:main"

[tool.setuptools]
packages = ["mcp_server"]
py-modules = ["cli"]
